Refactor with Ui, Storage, TaskList, Parser class
