--- src/main/java/action/Action.java
@@ -1,4 +1,0 @@
-package action;
-
-public abstract class Action {
-}
