--- src/main/java/action/ByeAction.java
@@ -1,11 +1,0 @@
-package action;
-
-public final class ByeAction extends Action {
-    public ByeAction() {
-    }
-
-    @Override
-    public String toString() {
-        return "bye";
-    }
-}
