--- src/main/java/action/ListAction.java
@@ -1,11 +1,0 @@
-package action;
-
-public final class ListAction extends Action {
-    public ListAction() {
-    }
-
-    @Override
-    public String toString() {
-        return "list";
-    }
-}
