--- src/main/java/action/UnmarkAction.java
@@ -1,18 +1,0 @@
-package action;
-
-public final class UnmarkAction extends Action {
-    private final int taskId;
-
-    public UnmarkAction(int taskId) {
-        this.taskId = taskId;
-    }
-
-    public int getTaskId() {
-        return taskId;
-    }
-
-    @Override
-    public String toString() {
-        return String.format("unmark %d", taskId);
-    }
-}
