+++ src/main/java/ekud/command/ByeCommand.java
@@ -1,0 +1,11 @@
+package ekud.command;
+
+public final class ByeCommand extends Command {
+    public ByeCommand() {
+    }
+
+    @Override
+    public String toString() {
+        return "bye";
+    }
+}
