+++ src/main/java/ekud/command/Command.java
@@ -1,0 +1,4 @@
+package ekud.command;
+
+public abstract class Command {
+}
