+++ src/main/java/ekud/command/ListCommand.java
@@ -1,0 +1,11 @@
+package ekud.command;
+
+public final class ListCommand extends Command {
+    public ListCommand() {
+    }
+
+    @Override
+    public String toString() {
+        return "list";
+    }
+}
