+++ src/main/java/ekud/command/MarkCommand.java
@@ -1,0 +1,18 @@
+package ekud.command;
+
+public final class MarkCommand extends Command {
+    private final int taskId;
+
+    public MarkCommand(int taskId) {
+        this.taskId = taskId;
+    }
+
+    public int getTaskId() {
+        return taskId;
+    }
+
+    @Override
+    public String toString() {
+        return String.format("mark %d", taskId);
+    }
+}
