+++ src/main/java/ekud/error/ArgumentException.java
@@ -1,0 +1,7 @@
+package ekud.error;
+
+public final class ArgumentException extends EkudException {
+    public ArgumentException(String message) {
+        super(message);
+    }
+}
