+++ src/main/java/ekud/error/EkudException.java
@@ -1,0 +1,7 @@
+package ekud.error;
+
+public class EkudException extends RuntimeException {
+    public EkudException(String message) {
+        super(message);
+    }
+}
