+++ src/main/java/ekud/error/StorageException.java
@@ -1,0 +1,7 @@
+package ekud.error;
+
+public final class StorageException extends EkudException {
+    public StorageException(String message) {
+        super(message);
+    }
+}
