+++ src/main/java/ekud/state/State.java
@@ -1,0 +1,13 @@
+package ekud.state;
+
+public final class State {
+    private final TaskList taskList;
+
+    public State() {
+        taskList = new TaskList();
+    }
+
+    public TaskList getTaskList() {
+        return taskList;
+    }
+}
