+++ src/main/java/ekud/state/Task.java
@@ -1,0 +1,33 @@
+package ekud.state;
+
+public abstract class Task {
+    private final String title;
+    private boolean isDone;
+
+    protected Task(String title, boolean isDone) {
+        this.title = title;
+        this.isDone = isDone;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public boolean isDone() {
+        return isDone;
+    }
+
+    public void mark() {
+        isDone = true;
+    }
+
+    public void unmark() {
+        isDone = false;
+    }
+
+    @Override
+    public String toString() {
+        String mark = isDone ? "X" : " ";
+        return "[" + mark + "] " + title;
+    }
+}
