+++ src/main/java/ekud/state/TodoTask.java
@@ -1,0 +1,12 @@
+package ekud.state;
+
+public class TodoTask extends Task {
+    public TodoTask(String title, boolean isDone) {
+        super(title, isDone);
+    }
+
+    @Override
+    public String toString() {
+        return "[T] " + super.toString();
+    }
+}
