+++ src/main/java/ekud/util/Pair.java
@@ -1,0 +1,19 @@
+package ekud.util;
+
+public final class Pair<T, U> {
+    private final T first;
+    private final U second;
+
+    public Pair(T first, U second) {
+        this.first = first;
+        this.second = second;
+    }
+
+    public T getFirst() {
+        return first;
+    }
+
+    public U getSecond() {
+        return second;
+    }
+}
