--- src/main/java/error/ArgumentException.java
@@ -1,7 +1,0 @@
-package error;
-
-public final class ArgumentException extends RuntimeException {
-    public ArgumentException(String message) {
-        super(message);
-    }
-}
