--- src/main/java/error/ParseException.java
@@ -1,14 +1,0 @@
-package error;
-
-public final class ParseException extends RuntimeException {
-    private final String line;
-
-    public ParseException(String line, String message) {
-        super(message);
-        this.line = line;
-    }
-
-    public String getLine() {
-        return line;
-    }
-}
