--- src/main/java/state/TodoTask.java
@@ -1,22 +1,0 @@
-package state;
-
-public class TodoTask extends Task {
-    public TodoTask(String title, boolean isDone) {
-        super(title, isDone);
-    }
-
-    @Override
-    public TodoTask mark() {
-        return new TodoTask(getTitle(), true);
-    }
-
-    @Override
-    public TodoTask unmark() {
-        return new TodoTask(getTitle(), false);
-    }
-
-    @Override
-    public String toString() {
-        return "[T] " + super.toString();
-    }
-}
